refactor(cli): split history.ts into focused conversation/generated-files modules
history.ts mixed three concerns in 816 LOC: history listing/CRUD + the
public formatters (kept here), conversation preview/transcript formatting
(→ history/conversationFormat.ts), and provider generated-file discovery
(→ history/generatedFiles.ts). Pure move: the public API and every
`@cli/runtime/history` import are unchanged; shared types stay in
history.ts and the new modules reference them via `import type` so there
is no runtime cycle.
