+++ packages/cli/src/runtime/history.ts
@@ -4,7 +4,6 @@ import {
   deleteAllExecutions,
   deleteExecution,
   getExecutionStore,
-  listExecutionWorkspaceFiles,
   listExecutions,
   type ExecutionListingEntry,
   type ExecutionMeta,
@@ -12,42 +11,26 @@ import {
 } from '@agent/storage';
 import type { AgentConfig } from '@agent/core/definition/AgentConfig';
 import type { CliNdjsonRecord } from '@cli/schemas/cliOutput';
-import { isFileNotFoundError } from '@common/errors';
 import {
   EXECUTION_STATUS,
   ExecutionIdSchema,
   type ExecutionId,
 } from '@shared/schemas';
 import { GoalStore } from '@tools/goal';
-import { StorageFS } from '@utils/files';
-import { byStringProp } from '@utils/core/comparators';
-import { isObject } from '@utils/core/typeGuards';
-import { isDirectory } from '@utils/files/fsEntryType';
-import { resolveStoragePath } from '@utils/files/taskRunStorage';
 
 import { readCliToolUseResumeDataForListing } from './toolUseResumeData';
 import { formatCliHistoryAgentLabel } from './historyLabels';
-
-const HISTORY_FILE_SCAN_DEPTH = 2;
-const CONVERSATION_PREVIEW_MESSAGE_LIMIT = 3;
-const CONVERSATION_PREVIEW_CONTENT_LIMIT = 4000;
-const HIDDEN_PROVIDER_REASONING_MARKER = '[provider reasoning hidden]';
-const WORKSPACE_FILE_TOOL_NAMES = new Set(['write_file', 'edit_file']);
-
-const KV_FILES = new Set([
-  'meta.json',
-  'config.json',
-  'conversation.json',
-  'todos.json',
-  'report.json',
-  'workspace-files.json',
-  'result-meta.json',
-]);
-
-interface ConversationMessageFormatOptions {
-  readonly includeToolUseMarkers?: boolean;
-  readonly contentLimit?: number;
-}
+import {
+  createConversationPreview,
+  createConversationTranscript,
+  formatConversationPreview,
+  formatConversationTranscript,
+} from './history/conversationFormat';
+import {
+  listGeneratedFiles,
+  listWorkspaceToolFiles,
+  mergeHistoryFiles,
+} from './history/generatedFiles';
 
 export interface CliHistoryEntry {
   readonly id: ExecutionId;
@@ -410,407 +393,4 @@ function teamPresetId(config: AgentConfig | null): string | undefined {
 function firstInputBasename(config: AgentConfig | null): string {
   const first = config?.inputFiles.at(0) ?? '';
   return first ? path.basename(first) : '-';
-}
-
-function isHistoryKvFile(name: string): boolean {
-  return (
-    KV_FILES.has(name) || name.startsWith('child-') || name.startsWith('flow_')
-  );
-}
-
-function createConversationPreview(
-  conversation: readonly unknown[] | null,
-): CliHistoryConversationPreview | null {
-  const transcript = buildConversationMessages(conversation, {
-    includeToolUseMarkers: false,
-    contentLimit: CONVERSATION_PREVIEW_CONTENT_LIMIT,
-  });
-  if (!transcript) return null;
-
-  const lastAssistant = transcript.messages.findLast((message) =>
-    isAssistantMessageRole(message.role),
-  );
-  const selected = lastAssistant
-    ? [lastAssistant]
-    : transcript.messages.slice(-CONVERSATION_PREVIEW_MESSAGE_LIMIT);
-
-  return {
-    messageCount: transcript.messageCount,
-    messages: selected,
-  };
-}
-
-function createConversationTranscript(
-  conversation: readonly unknown[] | null,
-): CliHistoryConversationPreview | null {
-  return buildConversationMessages(conversation, {
-    includeToolUseMarkers: true,
-  });
-}
-
-function buildConversationMessages(
-  conversation: readonly unknown[] | null,
-  options: ConversationMessageFormatOptions,
-): CliHistoryConversationPreview | null {
-  if (!conversation?.length) return null;
-  const messages = conversation
-    .map((message, i) => toConversationPreviewMessage(message, i + 1, options))
-    .filter((message) => message.content.trim().length > 0);
-  if (!messages.length) return null;
-  return {
-    messageCount: conversation.length,
-    messages,
-  };
-}
-
-function toConversationPreviewMessage(
-  message: unknown,
-  index: number,
-  options: ConversationMessageFormatOptions,
-): CliHistoryConversationPreviewMessage {
-  const raw = isObject(message) ? message : {};
-  const role = typeof raw.role === 'string' ? raw.role : 'unknown';
-  const content = formatConversationMessage(raw, options);
-  const truncated =
-    options.contentLimit !== undefined && content.length > options.contentLimit;
-  return {
-    index,
-    role,
-    content: truncated
-      ? `${content.slice(0, options.contentLimit).trimEnd()}\n...[truncated]`
-      : content,
-    truncated,
-  };
-}
-
-function formatConversationMessage(
-  raw: Record<string, unknown>,
-  options: ConversationMessageFormatOptions,
-): string {
-  const role = typeof raw.role === 'string' ? raw.role : '';
-  const parts = [
-    formatConversationMessageContent(raw.content, options),
-    formatConversationMessageContent(raw.parts, options),
-    ...(options.includeToolUseMarkers === true
-      ? formatTopLevelToolCalls(raw.tool_calls)
-      : []),
-  ].filter((part) => part.trim().length > 0);
-  if (parts.length > 0) return parts.join('\n').trim();
-  if (
-    isAssistantMessageRole(role) &&
-    (hasProviderReasoningBlock(raw.content) ||
-      hasProviderReasoningBlock(raw.parts))
-  ) {
-    return HIDDEN_PROVIDER_REASONING_MARKER;
-  }
-  return '';
-}
-
-function isAssistantMessageRole(role: string): boolean {
-  return role === 'assistant' || role === 'model';
-}
-
-function formatConversationPreview(
-  preview: CliHistoryConversationPreview,
-): string {
-  const shown =
-    preview.messages.length === 1
-      ? `${preview.messages[0]?.role ?? 'message'} message ${preview.messages[0]?.index ?? '?'}`
-      : `${preview.messages.length} recent messages`;
-  return formatConversationMessages(preview, shown);
-}
-
-function formatConversationTranscript(
-  transcript: CliHistoryConversationPreview,
-): string {
-  const shown =
-    transcript.messages.length === transcript.messageCount
-      ? 'all messages'
-      : `${transcript.messages.length} non-empty messages`;
-  return formatConversationMessages(transcript, shown);
-}
-
-function formatConversationMessages(
-  transcript: CliHistoryConversationPreview,
-  shown: string,
-): string {
-  const lines = [
-    `Conversation (${transcript.messageCount} messages; showing ${shown}):`,
-  ];
-  for (const message of transcript.messages) {
-    lines.push('', `[${message.role} #${message.index}]`, message.content);
-  }
-  return lines.join('\n');
-}
-
-function formatConversationMessageContent(
-  content: unknown,
-  options: ConversationMessageFormatOptions,
-): string {
-  if (content == null) return '';
-  if (typeof content === 'string') return content;
-  if (Array.isArray(content)) {
-    return content
-      .map((block) => formatConversationContentBlock(block, options))
-      .join('\n')
-      .trim();
-  }
-  return formatJsonish(content);
-}
-
-function formatConversationContentBlock(
-  block: unknown,
-  options: ConversationMessageFormatOptions,
-): string {
-  if (typeof block === 'string') return block;
-  if (!isObject(block)) return formatJsonish(block);
-  if (typeof block.text === 'string') return block.text;
-
-  if (isObject(block.functionCall)) {
-    if (options.includeToolUseMarkers !== true) return '';
-    const name =
-      typeof block.functionCall.name === 'string'
-        ? block.functionCall.name
-        : 'unknown';
-    return `[tool_use: ${name}]`;
-  }
-
-  if (isObject(block.functionResponse)) {
-    return `[tool_result: ${formatGoogleFunctionResponse(block.functionResponse, options)}]`;
-  }
-
-  switch (block.type) {
-    case 'thinking':
-    case 'redacted_thinking':
-      return '';
-    case 'tool_use':
-      if (options.includeToolUseMarkers !== true) return '';
-      return `[tool_use: ${String(block.name ?? 'unknown')}]`;
-    case 'tool_result':
-      return `[tool_result: ${formatConversationMessageContent(block.content, options)}]`;
-    default:
-      return formatJsonish(block);
-  }
-}
-
-function hasProviderReasoningBlock(content: unknown): boolean {
-  if (Array.isArray(content)) return content.some(isProviderReasoningBlock);
-  return isProviderReasoningBlock(content);
-}
-
-function isProviderReasoningBlock(block: unknown): boolean {
-  return (
-    isObject(block) &&
-    (block.type === 'thinking' || block.type === 'redacted_thinking')
-  );
-}
-
-function formatGoogleFunctionResponse(
-  functionResponse: Record<string, unknown>,
-  options: ConversationMessageFormatOptions,
-): string {
-  const response = isObject(functionResponse.response)
-    ? functionResponse.response
-    : undefined;
-  if (response && Object.hasOwn(response, 'result')) {
-    return formatConversationMessageContent(response.result, options);
-  }
-  if (response !== undefined) {
-    return formatConversationMessageContent(response, options);
-  }
-  return formatJsonish(functionResponse);
-}
-
-function formatTopLevelToolCalls(toolCalls: unknown): string[] {
-  if (!Array.isArray(toolCalls)) return [];
-  return toolCalls.map(formatTopLevelToolCall);
-}
-
-function formatTopLevelToolCall(toolCall: unknown): string {
-  if (!isObject(toolCall)) return `[tool_use: ${formatJsonish(toolCall)}]`;
-  const nestedFunction = isObject(toolCall.function)
-    ? toolCall.function
-    : undefined;
-  const name =
-    typeof nestedFunction?.name === 'string'
-      ? nestedFunction.name
-      : typeof toolCall.name === 'string'
-        ? toolCall.name
-        : 'unknown';
-  return `[tool_use: ${name}]`;
-}
-
-function formatJsonish(value: unknown): string {
-  try {
-    return JSON.stringify(value) ?? '';
-  } catch {
-    return String(value);
-  }
-}
-
-async function listGeneratedFiles(id: ExecutionId): Promise<CliHistoryFile[]> {
-  const runDir = await resolveStoragePath(id);
-  if (!runDir) return [];
-  return walkStorageDirectory(runDir, '', HISTORY_FILE_SCAN_DEPTH);
-}
-
-async function walkStorageDirectory(
-  basePath: string,
-  relativePath: string,
-  maxDepth: number,
-): Promise<CliHistoryFile[]> {
-  const fullPath = relativePath ? path.join(basePath, relativePath) : basePath;
-  const entries = await StorageFS.readDir(fullPath).catch((error: unknown) => {
-    if (isFileNotFoundError(error)) return [];
-    throw error;
-  });
-
-  const files: CliHistoryFile[] = [];
-  for (const [name, type] of entries) {
-    if (isHistoryKvFile(name)) continue;
-    const rawRelative = relativePath ? path.join(relativePath, name) : name;
-    const childPath = path.join(basePath, rawRelative);
-    const entryIsDirectory = isDirectory(type);
-    const stat = await StorageFS.stat(childPath).catch(() => ({ size: 0 }));
-    files.push({
-      path: rawRelative.replaceAll('\\', '/'),
-      size: stat.size,
-      isDirectory: entryIsDirectory,
-    });
-    if (entryIsDirectory && maxDepth > 1) {
-      files.push(
-        ...(await walkStorageDirectory(basePath, rawRelative, maxDepth - 1)),
-      );
-    }
-  }
-  return files.sort(byStringProp((f) => f.path));
-}
-
-async function listWorkspaceToolFiles(
-  config: AgentConfig | null,
-  persistedPaths: readonly string[],
-  conversation: readonly unknown[] | null,
-): Promise<CliHistoryFile[]> {
-  const filePaths = persistedPaths.length
-    ? persistedPaths
-    : conversation?.length
-      ? extractWorkspaceFileToolPaths(conversation)
-      : [];
-  const workspaceFiles = await listExecutionWorkspaceFiles(config, filePaths);
-  return workspaceFiles.map((file) => ({
-    path: file.displayPath,
-    size: file.size,
-    isDirectory: file.isDirectory,
-  }));
-}
-
-function extractWorkspaceFileToolPaths(
-  conversation: readonly unknown[],
-): string[] {
-  const paths: string[] = [];
-  for (const message of conversation) {
-    if (!isObject(message)) continue;
-
-    const responseToolPath = extractResponseFunctionCallFilePath(message);
-    if (responseToolPath) paths.push(responseToolPath);
-
-    const toolCalls = Array.isArray(message.tool_calls)
-      ? message.tool_calls
-      : [];
-    for (const toolCall of toolCalls) {
-      const toolPath = extractOpenAiToolCallFilePath(toolCall);
-      if (toolPath) paths.push(toolPath);
-    }
-
-    const contentBlocks = Array.isArray(message.content) ? message.content : [];
-    for (const block of contentBlocks) {
-      const toolPath = extractContentToolUseFilePath(block);
-      if (toolPath) paths.push(toolPath);
-    }
-
-    const parts = Array.isArray(message.parts) ? message.parts : [];
-    for (const part of parts) {
-      const toolPath = extractGoogleFunctionCallFilePath(part);
-      if (toolPath) paths.push(toolPath);
-    }
-  }
-  return paths;
-}
-
-function extractResponseFunctionCallFilePath(
-  message: Record<string, unknown>,
-): string | undefined {
-  if (message.type !== 'function_call') return undefined;
-  if (
-    typeof message.name !== 'string' ||
-    !WORKSPACE_FILE_TOOL_NAMES.has(message.name)
-  ) {
-    return undefined;
-  }
-  return extractToolArgumentsFilePath(message.arguments);
-}
-
-function extractOpenAiToolCallFilePath(toolCall: unknown): string | undefined {
-  if (!isObject(toolCall)) return undefined;
-  const fn = isObject(toolCall.function) ? toolCall.function : {};
-  if (typeof fn.name !== 'string' || !WORKSPACE_FILE_TOOL_NAMES.has(fn.name)) {
-    return undefined;
-  }
-  return extractToolArgumentsFilePath(fn.arguments);
-}
-
-function extractContentToolUseFilePath(block: unknown): string | undefined {
-  if (!isObject(block) || block.type !== 'tool_use') return undefined;
-  if (
-    typeof block.name !== 'string' ||
-    !WORKSPACE_FILE_TOOL_NAMES.has(block.name)
-  ) {
-    return undefined;
-  }
-  return extractToolArgumentsFilePath(block.input);
-}
-
-function extractGoogleFunctionCallFilePath(part: unknown): string | undefined {
-  if (!isObject(part) || !isObject(part.functionCall)) return undefined;
-  const { functionCall } = part;
-  if (
-    typeof functionCall.name !== 'string' ||
-    !WORKSPACE_FILE_TOOL_NAMES.has(functionCall.name)
-  ) {
-    return undefined;
-  }
-  return extractToolArgumentsFilePath(functionCall.args);
-}
-
-function extractToolArgumentsFilePath(
-  argumentsValue: unknown,
-): string | undefined {
-  const args = parseToolArguments(argumentsValue);
-  const toolPath = typeof args?.path === 'string' ? args.path.trim() : '';
-  return toolPath || undefined;
-}
-
-function parseToolArguments(
-  argumentsValue: unknown,
-): Record<string, unknown> | undefined {
-  if (isObject(argumentsValue)) return argumentsValue;
-  if (typeof argumentsValue !== 'string') return undefined;
-  try {
-    const parsed: unknown = JSON.parse(argumentsValue);
-    return isObject(parsed) ? parsed : undefined;
-  } catch {
-    return undefined;
-  }
-}
-
-function mergeHistoryFiles(
-  ...fileGroups: readonly (readonly CliHistoryFile[])[]
-): CliHistoryFile[] {
-  const files = new Map<string, CliHistoryFile>();
-  for (const group of fileGroups) {
-    for (const file of group) {
-      if (!files.has(file.path)) files.set(file.path, file);
-    }
-  }
-  return [...files.values()].sort(byStringProp((f) => f.path));
 }
